requirements.txt
aiohttp==3.6.2
alembic==1.0.11
amqp==2.5.1
aniso8601==7.0.0
astroid==2.2.5
async-timeout==3.0.1
asyncio==3.4.3
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.7
billiard==3.6.1.0
cached-property==1.5.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
Flask==1.1.1
Flask-RESTful==0.3.7
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.0
idna==2.7
importlib-metadata==0.19
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
kombu==4.6.5
lazy-object-proxy==1.4.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==2.19.5
mccabe==0.6.1
more-itertools==7.2.0
multidict==4.5.2
mypy-extensions==0.4.1
packaging==19.1
paramiko==2.6.0
pluggy==0.12.0
py==1.8.0
pycparser==2.19
pylint==2.3.1
PyNaCl==1.3.0
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.0.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
PyYAML==5.1
requests==2.20.1
six==1.12.0
SQLAlchemy==1.3.6
texttable==0.9.1
typed-ast==1.4.0
urllib3==1.24.3
vine==1.3.0
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.5
wrapt==1.11.2
yarl==1.3.0
zipp==0.5.2
python-jose = "*"